+++ src/Microsoft.AspNetCore.Mvc.ViewFeatures/ViewFeatures/DefaultHtmlGenerator.cs
@@ -13,6 +13,7 @@ using Microsoft.AspNetCore.Antiforgery;
 using Microsoft.AspNetCore.Html;
 using Microsoft.AspNetCore.Mvc.Internal;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
+using Microsoft.AspNetCore.Mvc.ModelBinding.Internal;
 using Microsoft.AspNetCore.Mvc.ModelBinding.Validation;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.AspNetCore.Mvc.Routing;
@@ -1441,8 +1442,8 @@ namespace Microsoft.AspNetCore.Mvc.ViewFeatures
         }
 
         private IHtmlContent GenerateGroupsAndOptions(
-            string optionLabel, 
-            IEnumerable<SelectListItem> selectList, 
+            string optionLabel,
+            IEnumerable<SelectListItem> selectList,
             ICollection<string> currentValues)
         {
             var listItemBuilder = new HtmlContentBuilder();
